qthink/requirements.txt
openai
python-dotenv
gitpython
faiss-cpu  # optional if you build vector memory
PyGithub
flask

